+++ ff_standard_lib/src/consolidators/heikinashi.rs
@@ -1,7 +1,7 @@
 use chrono::{DateTime, Duration, Utc};
 use crate::apis::vendor::client_requests::ClientSideDataVendor;
 use crate::consolidators::candlesticks::open_time;
-use crate::rolling_window::RollingWindow;
+use crate::standardized_types::rolling_window::RollingWindow;
 use crate::standardized_types::base_data::base_data_enum::BaseDataEnum;
 use crate::standardized_types::base_data::base_data_type::BaseDataType;
 use crate::standardized_types::enums::{Resolution, StrategyMode};
@@ -116,7 +116,7 @@ impl HeikinAshiConsolidator
 
 impl HeikinAshiConsolidator
 {
-    pub(crate) async fn new(subscription: DataSubscription, history_to_retain: usize) -> Result<HeikinAshiConsolidator, ConsolidatorError> {
+    pub(crate) async fn new(subscription: DataSubscription, history_to_retain: u64) -> Result<HeikinAshiConsolidator, ConsolidatorError> {
         if subscription.base_data_type != BaseDataType::Candles {
             return Err(ConsolidatorError { message: format!("{} is an Invalid base data type for HeikinAshiConsolidator", subscription.base_data_type) });
         }
@@ -141,7 +141,7 @@ impl HeikinAshiConsolidator
         })
     }
 
-    pub(crate) async fn new_and_warmup(subscription: DataSubscription, history_to_retain: usize, warm_up_to_time: DateTime<Utc>, strategy_mode: StrategyMode) -> Result<HeikinAshiConsolidator, ConsolidatorError> {
+    pub(crate) async fn new_and_warmup(subscription: DataSubscription, history_to_retain: u64, warm_up_to_time: DateTime<Utc>, strategy_mode: StrategyMode) -> Result<HeikinAshiConsolidator, ConsolidatorError> {
         if subscription.base_data_type != BaseDataType::Candles {
             return Err(ConsolidatorError { message: format!("{} is an Invalid base data type for HeikinAshiConsolidator", subscription.base_data_type) });
         }
