+++ ff_standard_lib/src/standardized_types/data_server_messaging.rs
@@ -172,7 +172,7 @@ pub enum SynchronousResponseType {
 
     TickSize(Symbol, f64),
     
-    DecimalAccuracy(Symbol, u32),
+    DecimalAccuracy(Symbol, u64),
 }
 
 impl SynchronousResponseType {
